Update chord_diagram.js
before all average_estimates of the new_data.csv were aggregated and displayed in the graph (chord_diagram.js). but that is not entirely correct, because some incidents of forced Migration are reported on by more than one row. For example we might have an Expulsion of 200000 People from 1989 to 1993 and one of 100000 from 1991 to 1992. the first entry covers 5 years (1993 is included) and the second two years. we assume a unifrom Distribution for both which makes as Thing that there are 40000 per year for the first row and 50000 for the second. that means that we have two year that have contradiciting numbers. now such such contradiction are found and then averaged. here is how we define such contradictions. two rows are considered to be reporting on the same Thing when the origin_name AND the target_name AND the expelled_group are identical AND there is at least Party overlapse in the time (start_year; stop_year).