Country Search Bar, Sankey Chart, Country Flags, Contradiction Estimate Uncertainty
New Functions:

Country Search Bar that allows to find countrys that are to small to be visible on the OD chart

Sankey Chart:
Allows focuse on singualr countries with allevial flow chart / Sankey chart

Aggregation logic:
New logic for aggregating diffrent estiamtes for the same event: when origin_country; target_coountry AND year are identical median is searched for (average of two medians for even numbers). min and max range are shown with lighter color in line_chart.